Move test util code into src folder, update package.json references, update test-util references across packages
