+++ packages/salesforcedx-apex-debugger/test/integration/adapter.test.ts
@@ -10,7 +10,7 @@ import {
   CommandExecution,
   SfdxCommandBuilder
 } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
-import * as util from '@salesforce/salesforcedx-test-utils-vscode/out/src/test/orgUtils';
+import * as util from '@salesforce/salesforcedx-test-utils-vscode/out/src/orgUtils';
 import { expect } from 'chai';
 import * as path from 'path';
 import * as rimraf from 'rimraf';
