+++ packages/salesforcedx-test-utils-vscode/package.json
@@ -2,14 +2,14 @@
   "name": "@salesforce/salesforcedx-test-utils-vscode",
   "displayName": "SFDX Test Utilities for VS Code",
   "description": "Provides test utilies to interface the SFDX libraries with VS Code",
-  "version": "45.2.0",
+  "version": "45.7.0",
   "publisher": "salesforce",
   "license": "BSD-3-Clause",
   "categories": [
     "Other"
   ],
   "dependencies": {
-    "@salesforce/salesforcedx-utils-vscode": "45.2.0",
+    "@salesforce/salesforcedx-utils-vscode": "45.7.0",
     "shelljs": "0.8.3"
   },
   "devDependencies": {
