+++ packages/salesforcedx-vscode-apex-debugger/package.json
@@ -29,7 +29,7 @@
     "vscode-extension-telemetry": "0.0.17"
   },
   "devDependencies": {
-    "@salesforce/salesforcedx-test-utils-vscode": "45.2.0",
+    "@salesforce/salesforcedx-test-utils-vscode": "45.7.0",
     "@types/chai": "^4.0.0",
     "@types/mocha": "2.2.38",
     "@types/node": "8.9.3",
