+++ packages/salesforcedx-vscode-lwc/package.json
@@ -35,7 +35,7 @@
     "vscode-languageclient": "^3.5.1"
   },
   "devDependencies": {
-    "@salesforce/salesforcedx-test-utils-vscode": "45.2.0",
+    "@salesforce/salesforcedx-test-utils-vscode": "45.7.0",
     "@types/chai": "^4.0.0",
     "@types/mocha": "2.2.38",
     "@types/node": "8.9.3",
