+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/index.ts
@@ -6,7 +6,7 @@
  */
 
 // tslint:disable-next-line:no-var-requires
-const testRunner = require('@salesforce/salesforcedx-test-utils-vscode/out/src/test/testrunner');
+const testRunner = require('@salesforce/salesforcedx-test-utils-vscode/out/src/testrunner');
 
 // You can directly control Mocha options by uncommenting the following lines
 // See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for more info
